Refactor to not have global side effects & state in script.js (#12)
To fix
https://github.com/ParkingReformNetwork/parking-lot-map/issues/10, I
want to add unit tests to test that we can handle several edge cases.
But our unit tests fail when importing `script.js` because at import
time, we do a bunch of side effects like using `L` (leaflet) and using
Jquery to modify the document.

Instead, to be able to import `script.js`, we can't have any "side
effects". All we should do when importing is define our functions.

Specifically, we have a new function `setUpSite()`, which will initiate
all the side effects when called. Our HTML then does this:

```html
    <script type="text/javascript" src="js/script.js"></script>
    <script>
      setUpSite();
    </script>
```

As a result, we can now add a simple unit test to confirm the import
works!

This is also a good refactor. It removes all global state from
`script.js`. Functions can only access what they specifically are passed
in as arguments. For example, it's now obvious where exactly we are
modifying the `map`.