Test and refactor the share link feature (#13)
To fix
https://github.com/ParkingReformNetwork/parking-lot-map/issues/10, I'll
need to make a structural change to `setUpCitiesLayer`.

Before doing that, I want to have better test coverage and clearer code.
I split this up a precursor PR to keep code review simpler.

This changes how we calculate the share link if the URL already has a
`#`. Currently, the only use of a `#` on our site is for setting a
pre-selected city. So, replacing the `#` entirely with the new value
makes the code much simpler.