Fix loading from a share link not working (#15)
Close https://github.com/ParkingReformNetwork/parking-lot-map/issues/10.

To simplify our support for loading from a share link, the `<option>`
values for the city `<select>` are now the normalized city names, like
`anchorage` and `fort worth`. This is not a future-proof scheme, per
https://github.com/ParkingReformNetwork/parking-lot-map/issues/10, but
it can be improved in a follow up.

We no longer use jQuery in this specific part of the codebase. jQuery is
no longer necessary thanks to improvements in native JavaScript the past
15 years. Native JavaScript has much better performance.