Label parking lots data set with city names (#20)
This will make it much easier to understand and edit the data set.

I used the script `label_cities.py` to automate this. It's deleted in
https://github.com/ParkingReformNetwork/parking-lot-map/pull/20/commits/71ae4db0100a72444507d76cb796820b6395b333
though to avoid having "dead code" in the repo, since this is a one-time
operation.

The algorithm was:
1. For each city, find its latitude and longitude pairs with the
precision of one decimal, e.g. `(89.1, 120.3)`. I validated with code
that no city is overlapping.
2. For each parking lot data set, find all of its unique lat/long pairs
with one decimal precision.
3. Look up which city those pairs correspond to. Ensure that every
single coordinate is contained within the city boundaries.

This results in some floats losing precision because Python's `float`
representation is not as precise. This should not matter; the new float
value is still incredibly precise.