JavaScript stringify parking-lots.geojson (#26)
We'll soon run a script that dynamically updates `parking-lots.geojson`
using JavaScript. When it does that, it formats things differently than
now.

To make the review easier, this PR simply uses JavaScript to read in the
file, then immediately writes it back using `JSON.stringify()`. There
should be no meaningful changes to the data.