Make usage of global objects explicit (#23)
JavaScript defines a couple global variables like `document` and
`window.location.href`.

But code is much easier to understand and test when it only relies on
its explicit arguments and does not access global variables/state. It
allows the reader to know immediately from the function's signature what
it might do.

For example, now it's clear which functions modify the global `document`
(DOM) and which cannot possibly do so. Whereas before, you had to read
the entire function body.

We use ESLint to enforce that globals are only used in `setUpSite()`. It
then passes those global variables explicitly to all the other
functions.