Reorganize project (#36)
Reorganizes into an `assets` and `src` folder. Within `src/js` and
`src/css`, we have a `vendor` folder to make it more clear what is our
own code. `data/`, `scripts/`, and `__tests__` stay the same.

```
src
├── css
│   ├── style.css
│   └── vendor
│       ├── bootstrap.css
│       └── leaflet.zoomhome.css
└── js
    ├── script.js
    └── vendor
        └── leaflet.zoomhome.js
assets
├── fonts
│   ├── glyphicons-halflings-regular.eot
│   ├── glyphicons-halflings-regular.svg
│   ├── glyphicons-halflings-regular.ttf
│   ├── glyphicons-halflings-regular.woff
│   └── glyphicons-halflings-regular.woff2
└── images
    ├── parking-reform-network.png
    └── share-url-button.png
```

The project organization doesn't matter much because Parcel will rewrite
all the imports for us. So, we can go with what is easiest to maintain.

I wanted to move `index.html` into `src/`, but that causes issues with
`parcel build`. So I'll hold off on that change until maybe a follow up.