Detect the geometry type automatically when updating data (#45)
The `update-city-boundaries` script failed when uploading the Birmingham
data because it used `StraightLine` rather than `Polygon` like past
entries. That caused Leaflet to fail to process the GeoJSON.

It doesn't matter which we use, as long as we respect the original data
source. We already validate that there is only one `Feature`, which does
matter.