Add `serve-dist` and `test-dist` commands (#60)
This is prework for continuous deployment. 

We'll have the deploy workflow run on every push to `main`. Before a
deploy, we want to confirm that the `dist/` folder works. That's a more
relevant test than the source code working.

So, `serve-dist` will use `http-server` to start up the `dist/` folder.
Then, `test-dist` will run the integration tests, but on port 8080
rather than 1234.