Remove "Select a city" option from city toggle (#73)
This option was confusing once the site loads. It didn't actually work,
and its semantics are unclear anyways.

It's somewhat useful when the site is first loading, but we set up the
city toggle so fast (especially after
https://github.com/ParkingReformNetwork/parking-lot-map/pull/71) that it
doesn't matter. It's fine to show an empty select box for a couple of
milliseconds.