Stop forcing links to open in new tabs (#48)
We were setting `target="_blank"` on each link, which forces the link to
open in a new tab/window when the user clicks on it.

This is seen as not accessible to certain users, especially those using
assistive technologies like screen readeres. It also removes the choice
from users what they want to do.