Stop using `L` global (#76)
This makes the code easier to understand and is more aligned with modern
JavaScript.

The Leaflet devs have also announced that Leaflet 2.0 will stop having
the `L` global, so this is prework.

Theoretically, using more precise imports allows Parcel to use "tree
shaking" to only include the JavaScript we actually use. The bundle size
is the same before and after this PR, so it doesn't look like it's
happening. But maybe with Leaflet 2.0 they will redesign their code to
allow that to happen.