Remove zoom to home button (#111)
This was the globe button, which would take you to a view of the whole
US.
<img width="60" alt="Captura de pantalla 2023-08-10 a la(s) 3 23 51 p m"
src="https://github.com/ParkingReformNetwork/parking-lot-map/assets/14852634/c49fbfc8-e1e6-40ee-b03d-64a775158912">

We decided that this is a bad user experience because the map is not
very useful when zoomed out far. The important user experience with the
map is the city detail, to see the actual lots themselves. There is
nothing interesting about simply seeing the # of cities we have mapped.

Further, our user experience is not optimized to let you zoom into a
city: https://github.com/ParkingReformNetwork/parking-lot-map/issues/79.
We think this would be non-trivial to do and is not important since
users have the city toggle in the corner already to change views.

So, this was a misfeature that directed users to a UX we aren't
optimized for. It also added code complexity.