Refactor updating scorecard (#169)
Why this refactor?
This simplifies when scorecard updates. 

Before, there were many different cases in which a scorecard would
update, like the dropdown changing, the user panning to a different
city, or the user clicking on a city.
Now, the scorecard only updates to the city closes to the center of the
screen. So, when the dropdown changes, the map snaps to city, which will
also automatically update the scorecard.

---------

Co-authored-by: tunglinn <tunglin@yahoo.com>