Define breakpoints and font sizes (#198)
https://www.refactoringui.com strongly recommends using a design system
where we systematize colors, font sizes, and breakpoints. That is all
moved to a new `theme/` folder. This will be useful with the mandates
map, that we can keep in sync the same files.

The breakpoint sizes come from tailwind.css.

The font scale comes from refactoringui.com. It strongly recommends
using `px` over `em`, and generally `px` over `rem` but `em` is an
issue.