Use `calc()` in Sass (#201)
One issue with our styling is that it's pretty haphazard. It's never
clear why certain values are chosen, and we often repeat values too.

By using `calc()`, we can make the code more self-documenting and
flexible. When you change a variable, it should correctly update all the
other styling code automatically.

--

This also recalibrates the icon sizes & removes the icon mixin, which
didn't end up working well with the project.

--

This PR should have no impact on the end user. It is only a refactor.