Use flexbox for scorecard header (#207)
Even though I'll be proposing moving the share link icon to the top
header, it will still be in the scorecard as part of the initial
accordion revamp. So it's important that the styling is right.

This uses flexbox so that the share link icon is always positioned in
the correct spot, whereas using `float` was haphazard.

It also allows us to shrink the width of the scorecard.