Add `Observable.initialize()` (#255)
Before, you had to explicitly call every subscriber with the initial
state. Now, you can call `myObservable.initialize()` to do that.

Beyond DRY, this is helpful to encapsulate who the subscribers are. For
example, for the selectedCity observable, we'll have subscribers from
several files and we don't want to have to couple them all together.