Open all links in a new tab (#268)
The app is usually accessed via an iframe. If you click links currently,
the new page is in the iframe itself. This is really confusing. We could
use `target=_top` to replace the current page, but Tony has said in the
past that we should go to a new tab.

We now consistently use the up-right arrow icon with every external link
to make more explicit that the links go to a new tab.

<img width="792" alt="Screenshot 2024-08-02 at 8 50 24 AM"
src="https://github.com/user-attachments/assets/7640e41b-b4f3-4519-9d60-34a7e8f730f9">