Support new Connecticut city IDs (#302)
The CT map will have entries like `Hartford - rail station`, which
should convert to `hartford-rail-station`. We can do that with Lodash's
kebabCase.

That results in changing `st.` to become `st`, which is a good change
anyways. We ensure we still pre-load the correct place if someone uses
an anchor with the old style of `st.`.